Move around where we configure the metrics
